+++ src/win32/w32_stack.h
@@ -116,7 +116,7 @@ int git_win32__stack_compare(
  * @param suffix String written after each frame; defaults to "\n".
  */
 int git_win32__stack_format(
-	char *pbuf, int buf_len,
+	char *pbuf, size_t buf_len,
 	const git_win32__stack__raw_data *pdata,
 	const char *prefix, const char *suffix);
 
@@ -132,7 +132,7 @@ int git_win32__stack_format(
  * @param suffix String written after each frame; defaults to "\n".
  */
 int git_win32__stack(
-	char * pbuf, int buf_len,
+	char * pbuf, size_t buf_len,
 	int skip,
 	const char *prefix, const char *suffix);
